# docs/openapi.yaml
openapi: 3.0.3
info:
  title: APP_ModularDeGestion API
  description: >
    Contrato de la API para el sistema de gestión modular multi-negocio.
    Fuente de verdad compartida entre frontend y backend — ver
    docs/ARCHITECTURE.md para el contexto de dominio.

    Auth: todo endpoint requiere `Authorization: Bearer <token>` (ver
    POST /auth/login), excepto /health y el login mismo. El negocio_id de
    cualquier ruta /negocios/{negocioId}/... se valida contra el negocio_id
    del token — no alcanza con mandar un negocio_id distinto en la URL para
    operar sobre otro negocio. Un usuario con rol admin no está atado a
    ningún negocio y puede operar sobre cualquiera; el resto de roles
    (hoy solo "dueño") solo sobre el suyo. Devuelve 401 sin token o con uno
    inválido/expirado, 403 con un token válido pero sin permiso sobre ese
    negocio_id o sin rol admin donde se requiere.
  version: 0.1.0
servers:
  - url: http://localhost:8000/api/v1
    description: Desarrollo local
...
security:
  - bearerAuth: []
paths:
...
  /negocios:
    get:
      tags: [negocios]
      summary: Listar negocios
      description: Admin-only. Un usuario de negocio no ve la lista completa.
      operationId: listNegocios
      responses:
        "200":
          description: Lista de negocios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Negocio"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [negocios]
      summary: Crear negocio
      description: >
        Admin-only. Crea el negocio y su primer usuario (con credenciales
        propias) en el mismo paso — es el endpoint que consume el panel de
        admin para dar de alta un negocio nuevo.
      operationId: createNegocio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NegocioCreate"
      responses:
        "201":
          description: Negocio creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Negocio"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /negocios/{negocioId}:
    parameters:
      - $ref: "#/components/parameters/negocioId"
    get:
      tags: [negocios]
      summary: Obtener un negocio
      operationId: getNegocio
      responses:
        "200":
          description: Negocio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Negocio"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [negocios]
      summary: Editar configuración del negocio
      description: >
        Update parcial (solo se tocan los campos presentes en el body).
        Pensado sobre todo para `link_sunat`, editable desde la pestaña
        SUNAT del frontend.
      operationId: updateNegocio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NegocioUpdate"
      responses:
        "200":
          description: Negocio actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Negocio"
        "404":
          $ref: "#/components/responses/NotFound"
  /negocios/{negocioId}/usuarios:
    parameters:
      - $ref: "#/components/parameters/negocioId"
    get:
      tags: [usuarios]
      summary: Listar usuarios del negocio
      description: >
        Cada negocio arranca con un único usuario, creado en el mismo paso
        que el negocio (`POST /negocios`, admin-only). Este endpoint sirve
        para sumar más usuarios al negocio, o para resolver el `usuario_id`
        que pide `POST /movimientos` sin hardcodearlo.
      operationId: listUsuarios
      responses:
        "200":
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Usuario"
    post:
      tags: [usuarios]
      summary: Crear usuario adicional
      operationId: createUsuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsuarioCreate"
      responses:
        "201":
          description: Usuario creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
        "404":
          $ref: "#/components/responses/NotFound"
  /negocios/{negocioId}/productos:
    parameters:
      - $ref: "#/components/parameters/negocioId"
    get:
      tags: [productos]
      summary: Listar productos del negocio
      operationId: listProductos
      responses:
        "200":
          description: Lista de productos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Producto"
    post:
      tags: [productos]
      summary: Crear producto
      description: >
        Regla transversal anti-duplicados: si el nombre se parece a uno ya
        existente en el negocio, responde 409 con los candidatos en vez de
        crear. El frontend reintenta con `confirmar_nuevo=true` si de
        verdad es un ítem distinto.
      operationId: createProducto
      parameters:
        - name: confirmar_nuevo
          in: query
          required: false
          schema:
            type: boolean
            default: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductoCreate"
      responses:
        "201":
          description: Producto creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Producto"
        "409":
          description: Existe(n) producto(s) con nombre parecido; ver `candidatos` en el detalle.
  /negocios/{negocioId}/productos/{productoId}/ajustar-stock:
    parameters:
      - $ref: "#/components/parameters/negocioId"
      - name: productoId
        in: path
        required: true
        schema:
          type: integer
    post:
      tags: [productos]
      summary: Sumar o restar stock sin costo ni impacto en SUNAT
      description: >
        Complementa a registro-compras: no lleva costo_unitario ni queda
        en el historial de compras, así que no está gateado por
        modulo_rus_activo — sirve para reponer sin factura a mano o
        corregir un conteo, en cualquier negocio.
      operationId: ajustarStockProducto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductoAjusteStock"
      responses:
        "200":
          description: Producto con el stock actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Producto"
        "400":
          description: delta es 0, o dejaría el stock en negativo
        "404":
          $ref: "#/components/responses/NotFound"
...
  /negocios/{negocioId}/registro-compras:
    parameters:
      - $ref: "#/components/parameters/negocioId"
    get:
      tags: [registro-compras]
      summary: Listar registros de compra/reposición (requiere módulo RUS activo)
      operationId: listRegistroCompras
      responses:
        "200":
          description: Lista de registros de compra
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RegistroCompra"
        "403":
          description: Módulo RUS no activo para este negocio
    post:
      tags: [registro-compras]
      summary: Registrar compra/reposición de stock
      description: Suma `cantidad` al stock_actual del producto indicado.
      operationId: createRegistroCompra
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistroCompraCreate"
      responses:
        "201":
          description: Registro de compra creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistroCompra"
        "400":
          description: cantidad inválida (debe ser mayor a 0)
        "403":
          description: Módulo RUS no activo para este negocio
        "404":
          $ref: "#/components/responses/NotFound"
  /negocios/{negocioId}/sync:
    parameters:
      - $ref: "#/components/parameters/negocioId"
    post:
      tags: [sync]
      summary: >
        Procesa la cola de sincronización offline enviada por el cliente
        (frontend). Idempotente por item vía su id generado en cliente.
      operationId: pushSync
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/ColaSyncItem"
      responses:
        "200":
          description: Resultado de la sincronización por item
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ColaSyncResultado"
components:
  parameters:
    negocioId:
      name: negocioId
      in: path
      required: true
      schema:
        type: integer
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    NotFound:
      description: No encontrado
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
    Unauthorized:
      description: Falta el token, o es inválido/expiró
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
    Forbidden:
      description: Token válido, pero sin permiso sobre este negocio (o falta rol admin)
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
  schemas:
...
    NegocioBase:
      type: object
      properties:
        nombre: { type: string }
        rubro: { type: string }
        modulos_activos:
          type: object
          additionalProperties: true
          description: Flags de módulos opcionales activos para este negocio.
        plan_estado: { type: string }
        fecha_ultimo_pago: { type: string, format: date, nullable: true }
        link_sunat: { type: string, nullable: true }
        modulo_rus_activo: { type: boolean }
      required: [nombre, rubro]
    Negocio:
      allOf:
        - $ref: "#/components/schemas/NegocioBase"
        - type: object
          properties:
            id: { type: integer }
          required: [id]
    NegocioCreate:
      allOf:
        - $ref: "#/components/schemas/NegocioBase"
        - type: object
          properties:
            usuario_inicial:
              allOf:
                - $ref: "#/components/schemas/UsuarioCreate"
              description: >
                Primer usuario del negocio (con sus credenciales). Sin esto
                el negocio quedaría creado pero inaccesible para siempre.
          required: [usuario_inicial]
    NegocioUpdate:
      type: object
      description: >
        Update parcial: todos los campos opcionales, solo se aplican los
        que vengan en el body (mismo criterio que ClienteVehiculoUpdate).
      properties:
        nombre: { type: string }
        rubro: { type: string }
        modulos_activos:
          type: object
          additionalProperties: true
        plan_estado: { type: string }
        fecha_ultimo_pago: { type: string, format: date, nullable: true }
        link_sunat: { type: string, nullable: true }
        modulo_rus_activo: { type: boolean }
    UsuarioBase:
      type: object
      properties:
        nombre: { type: string }
        rol:
          type: string
          description: '"admin" (sin negocio_id, cross-tenant) o cualquier otro valor atado a un negocio (hoy solo "dueño").'
        username: { type: string }
      required: [nombre, rol, username]
    UsuarioCreate:
      allOf:
        - $ref: "#/components/schemas/UsuarioBase"
        - type: object
          properties:
            password:
              type: string
              description: Plaintext solo de paso — el backend la hashea y nunca la devuelve.
          required: [password]
    Usuario:
      allOf:
        - $ref: "#/components/schemas/UsuarioBase"
        - type: object
          properties:
            id: { type: integer }
            negocio_id:
              type: integer
              nullable: true
              description: null si es admin.
          required: [id, negocio_id]
    ProductoBase:
      type: object
      properties:
        nombre: { type: string }
        medida: { type: string, nullable: true }
        marca: { type: string, nullable: true }
        estado_uso: { type: string, nullable: true }
        precio_lista: { type: string, format: decimal }
        precio_compra: { type: string, format: decimal }
        clasificacion: { type: string, nullable: true }
        stock_actual: { type: integer }
        stock_minimo: { type: integer }
        activo: { type: boolean }
      required: [nombre]
    Producto:
      allOf:
        - $ref: "#/components/schemas/ProductoBase"
        - type: object
          properties:
            id: { type: integer }
            negocio_id: { type: integer }
          required: [id, negocio_id]
    ProductoCreate:
      $ref: "#/components/schemas/ProductoBase"
      description: negocio_id no va en el body — viene del path /negocios/{negocioId}/productos.
    ProductoAjusteStock:
      type: object
      properties:
        delta:
          type: integer
          description: Positivo repone, negativo corrige. No puede ser 0.
      required: [delta]
...
    RegistroCompraCreate:
      type: object
      properties:
        producto_id: { type: integer }
        cantidad: { type: integer, minimum: 1 }
        costo_unitario: { type: string, format: decimal }
        fecha:
          type: string
          format: date
          nullable: true
          description: Si se omite, se usa la fecha del día.
      required: [producto_id, cantidad, costo_unitario]
    RegistroCompra:
      type: object
      properties:
        id: { type: integer }
        negocio_id: { type: integer }
        producto_id: { type: integer }
        cantidad: { type: integer }
        costo_unitario: { type: string, format: decimal }
        fecha: { type: string, format: date }
    ColaSyncItem:
      type: object
      description: Item generado por el cliente (frontend) mientras estuvo offline.
      properties:
        id:
          type: string
          description: id generado en cliente (uuid), usado para idempotencia.
        entidad: { type: string, description: "ej. movimiento, producto" }
        payload:
          type: object
          additionalProperties: true
        fecha_creacion: { type: string, format: date-time }
      required: [id, entidad, payload]
    ColaSyncResultado:
      type: object
      properties:
        id: { type: string }
        estado: { type: string, enum: [aplicado, error, duplicado] }
        detalle: { type: string, nullable: true }